+++ tests/web/47691_test.dart
@@ -1,0 +1,26 @@
+// Copyright (c) 2021, the Dart project authors.  Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
+import 'package:expect/expect.dart';
+
+class LoadElements {
+  List call() => [];
+}
+
+class ViewModel {
+  ViewModel(this._loadElements);
+
+  final LoadElements _loadElements;
+
+  void init() {
+    final elements = _loadElements();
+    for (final element in elements) {
+      Expect.identical(element, element);
+    }
+  }
+}
+
+void main() {
+  ViewModel(LoadElements()).init();
+}